en cours) file d'attente

+++ aquiletour_core/src/main/java/ca/aquiletour/core/pages/queue/messages/ModifyAppointmentDurations.java
@@ -5,13 +5,13 @@ import ca.ntro.messages.NtroUserMessage;
 
 public class ModifyAppointmentDurations extends NtroUserMessage<User> {
 	
-	private long durationIncrementSeconds;
+	private int durationIncrementSeconds;
 
-	public long getDurationIncrementSeconds() {
+	public int getDurationIncrementSeconds() {
 		return durationIncrementSeconds;
 	}
 
-	public void setDurationIncrementSeconds(long durationIncrementSeconds) {
+	public void setDurationIncrementSeconds(int durationIncrementSeconds) {
 		this.durationIncrementSeconds = durationIncrementSeconds;
 	}
 }
